+++ src/main/java/ladysnake/blast/common/network/Packets.java
@@ -6,11 +6,11 @@ import net.fabricmc.fabric.api.server.PlayerStream;
 import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.entity.Entity;
 import net.minecraft.network.Packet;
+import net.minecraft.network.PacketByteBuf;
 import net.minecraft.network.packet.s2c.play.CustomPayloadS2CPacket;
 import net.minecraft.network.packet.s2c.play.EntitySpawnS2CPacket;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.util.Identifier;
-import net.minecraft.util.PacketByteBuf;
 
 import java.io.IOException;
 import java.util.Optional;
@@ -33,7 +33,7 @@ public final class Packets {
         return new CustomPayloadS2CPacket(SPAWN, bytes);
     }
 
-    public static <T extends Packet<?>> Optional<T> readFrom(final PacketByteBuf bytes, final Supplier<T> packet) {
+    public static  <T extends Packet<?>> Optional<T> readFrom(final PacketByteBuf bytes, final Supplier<T> packet) {
         final T deserializedPacket = packet.get();
         try {
             deserializedPacket.read(bytes);